ml-backend/requirement.txt
fastapi
uvicorn[standard]
pydantic-settings 
python-dotenv
sqlalchemy         
pyodbc             
pandas    
pandas_ta         
yfinance           
scikit-learn
numpy
tensorflow
bcrypt
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-multipart